# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dns.LookupRequest:
    properties:
      domain:
        type: string
    required:
    - domain
    type: object
  dns.LookupResponse:
    properties:
      domain:
        type: string
      error:
        type: string
      ipv4:
        items:
          type: string
        type: array
      ipv6:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Backend API for the Mihombreng controller
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Mihombreng API
  version: "1.0"
paths:
  /app/config:
    get:
      consumes:
      - application/json
      description: Get current Mihombreng application configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get application configuration
      tags:
      - Mihombreng App
    put:
      consumes:
      - application/json
      description: Update Mihombreng application configuration
      parameters:
      - description: Configuration to update
        in: body
        name: config
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Update application configuration
      tags:
      - Mihombreng App
  /app/geo/ipv4:
    get:
      description: Get IPv4 address with geolocation information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv4 geolocation
      tags:
      - Mihombreng App
  /app/geo/ipv6:
    get:
      description: Get IPv6 address with geolocation information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv6 geolocation
      tags:
      - Mihombreng App
  /app/ipv4:
    get:
      description: Get current IPv4 address
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv4 address
      tags:
      - Mihombreng App
  /app/ipv6:
    get:
      description: Get current IPv6 address
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties: true
            type: object
      summary: Get IPv6 address
      tags:
      - Mihombreng App
  /backup/create:
    post:
      description: Create a tar.gz backup of the entire working directory
      produces:
      - application/gzip
      responses:
        "200":
          description: Backup file
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create backup
      tags:
      - Backup
  /backup/restore:
    post:
      consumes:
      - multipart/form-data
      description: Restore configuration from uploaded tar.gz backup file
      parameters:
      - description: Backup file (.tar.gz)
        in: formData
        name: backup
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Restore backup
      tags:
      - Backup
  /dns/lookup:
    post:
      consumes:
      - application/json
      description: Resolve domain name to IPv4 and IPv6 addresses
      parameters:
      - description: Domain to lookup
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dns.LookupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dns.LookupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dns.LookupResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dns.LookupResponse'
      summary: Lookup domain IP addresses
      tags:
      - DNS
  /mihomo/{dir}:
    get:
      description: Get list of all files in the specified directory
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of files
          schema:
            items:
              type: string
            type: array
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get files from directory
      tags:
      - Mihomo files
    post:
      consumes:
      - application/json
      description: Create a new file with the specified content
      parameters:
      - description: Directory name (configs, proxy_providers, rule_providers)
        in: path
        name: dir
        required: true
        type: string
      - description: Filename and content
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new file
      tags:
      - Mihomo files
...
  /mihomo/active-config:
    get:
      description: Get the currently active mihomo configuration file path
      produces:
      - application/json
      responses:
        "200":
          description: Active config path
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties: true
            type: object
      summary: Get active config path
      tags:
      - Mihomo files
    put:
      consumes:
      - application/json
      description: Set the active mihomo configuration file path
      parameters:
      - description: Filename to set as active
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error message
          schema:
            additionalProperties: true
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties: true
            type: object
      summary: Set active config path
      tags:
      - Mihomo files
...
  /mihomo/dashboard-info:
    get:
      description: Get external-controller port, secret, and available dashboards
        from ui directory
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard information
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error message
          schema:
            additionalProperties: true
            type: object
      summary: Get Mihomo dashboard information
      tags:
      - Mihomo
  /mihomo/restart:
    post:
      consumes:
      - application/json
      description: Restart the mihomo service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Restart mihomo service
      tags:
      - Mihomo
  /mihomo/start:
    post:
      consumes:
      - application/json
      description: Start the mihomo service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Start mihomo service
      tags:
      - Mihomo
...
  /mihomo/stop:
    post:
      consumes:
      - application/json
      description: Stop the mihomo service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Stop mihomo service
      tags:
      - Mihomo
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"